+++ packages/upload-api/src/upload/list.js
@@ -10,13 +10,6 @@ export function uploadListProvider(context) {
   return Server.provide(Upload.list, async ({ capability }) => {
     const { cursor, size, pre } = capability.nb
     const space = Server.DID.parse(capability.with).did()
-
-    return {
-      ok: await context.uploadTable.list(space, {
-        size,
-        cursor,
-        pre,
-      }),
-    }
+    return await context.uploadTable.list(space, { size, cursor, pre })
   })
 }