+++ packages/upload-client/src/types.ts
@@ -177,7 +177,7 @@ export interface CARMetadata extends CARHeaderInfo {
    *
    * @see https://github.com/filecoin-project/FIPs/pull/758/files
    */
-  piece?: PieceLink
+  piece: PieceLink
   /**
    * Size of the CAR file in bytes.
    */
@@ -271,8 +271,6 @@ export interface UploadOptions
     ShardStoringOptions,
     UploadProgressTrackable {
   onShardStored?: (meta: CARMetadata) => void
-  /** when true, uploading will calculate filecoin piece link */
-  piece?: true
 }
 
 export interface UploadDirectoryOptions