+++ packages/upload-client/test/store.test.js
@@ -34,6 +34,7 @@ describe('Store.add', () => {
       url: 'http://localhost:9200',
       link: car.cid,
       with: space.did(),
+      allocated: car.size,
     }
 
     const service = mockService({
@@ -105,6 +106,7 @@ describe('Store.add', () => {
       url: 'http://localhost:9400', // this bucket always returns a 400
       link: car.cid,
       with: space.did(),
+      allocated: car.size,
     }
 
     const service = mockService({
@@ -158,6 +160,7 @@ describe('Store.add', () => {
       url: 'http://localhost:9500', // this bucket always returns a 500
       link: car.cid,
       with: space.did(),
+      allocated: car.size,
     }
 
     const service = mockService({
@@ -256,6 +259,7 @@ describe('Store.add', () => {
       url: 'http://localhost:9200',
       link: car.cid,
       with: space.did(),
+      allocated: car.size,
     }
 
     const service = mockService({